requirements.txt
pyPDF2
pdfplumber
pandas
requests
flask
bandit
mongoengine
python-dotenv